Initialize and dispose notification repository
Updated the `ChatApp` to initialize the `NotificationRepository` when the user is authenticated and dispose of it when not.

Also, refactored the dependency injection setup to register the necessary data sources and the `NotificationRepository` itself.

Removed the `NotificationService` as it is no longer needed.

Adjusted the `main.dart` to use `DefaultFirebaseOptions` for Firebase initialization.