Add `isLinking` parameter to `verifyPhoneNumber` and `verifyOtpCode` methods
Added a new `isLinking` boolean parameter to the `verifyPhoneNumber` and `verifyOtpCode` methods in `AuthDataSource` and `FirebaseAuthDataSource` to handle the linking of phone numbers to existing user accounts.

This change allows the methods to differentiate between signing in with a phone number and linking a phone number to an existing user.