Add Google Services configuration files for dev and prod environments
This commit adds the `google-services.json` files for both development and production environments in the Android project.

It also updates the `firebase.json` file to reflect the correct project IDs and app IDs for the production environment.

Additionally, new `GoogleService-Info.plist` files are added for both dev and prod environments in the iOS project to ensure proper Firebase integration across different build configurations.